@@ -12,14 +12,14 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "17.2.4",
-    "@angular/common": "17.2.4",
-    "@angular/compiler": "17.2.4",
-    "@angular/core": "17.2.4",
-    "@angular/forms": "17.2.4",
-    "@angular/platform-browser": "17.2.4",
-    "@angular/platform-browser-dynamic": "17.2.4",
-    "@angular/router": "17.2.4",
+    "@angular/animations": "17.3.0",
+    "@angular/common": "17.3.0",
+    "@angular/compiler": "17.3.0",
+    "@angular/core": "17.3.0",
+    "@angular/forms": "17.3.0",
+    "@angular/platform-browser": "17.3.0",
+    "@angular/platform-browser-dynamic": "17.3.0",
+    "@angular/router": "17.3.0",
     "@fortawesome/fontawesome-free": "^6.5.1",
     "@nx/angular": "18.1.1",
     "ag-grid-angular": "^30.1.0",
@@ -40,8 +40,8 @@
     "@angular-eslint/eslint-plugin-template": "17.0.1",
     "@angular-eslint/template-parser": "17.0.1",
     "@angular/cli": "~17.2.0",
-    "@angular/compiler-cli": "17.2.4",
-    "@angular/language-service": "17.2.4",
+    "@angular/compiler-cli": "17.3.0",
+    "@angular/language-service": "17.3.0",
     "@nx/eslint": "18.1.1",
     "@nx/eslint-plugin": "18.1.1",
     "@nx/jest": "18.1.1",
